loader: move DeriveMasqIPAddrFromInterface to datapath orchestrator config
This removes yet another way of passing in configuration to the loader and
sets it up so all unexported members of *loader can be made free functions
in a follow-up commit.
